requirements/development.txt
-r common.txt # includes the base.txt requirements file
coverage==3.7.1
django-debug-toolbar==1.6.0
selenium==2.45.0
pytest==2.7.2
pytest-cov==2.1.0
pytest-django==2.8.0